requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
tensorflow>=2.12.0
streamlit>=1.28.0
plotly>=5.17.0
google-genai>=1.0.0

